prosjekt/requirements.txt
numpy
matplotlib.pyplot
seaborn
sklearn.metrics